requirements.txt
pyyaml
spotipy
pyvis
networkx
streamlit
pydantic